pyproject.toml
[tool.poetry]
name = "dev_log"
version = "0.1.0"
description = "blog made in django"
authors = ["hsw0905 <hsw0905@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
pip = "^21.0.1"
setuptools = "^54.1.1"
python-decouple = "^3.4"
psycopg2-binary = "^2.8.6"
django-storages = "^1.11.1"
django-environ = "^0.4.5"
dj-inmemorystorage = "^2.1.0"
boto3 = "^1.17.25"
Pillow = "^8.1.2"
gunicorn = "^20.0.4"
supervisor = "^4.2.2"
requests = "^2.25.1"
djangorestframework = "^3.12.2"
django-webpack-loader = "^0.7.0"
whitenoise = "^5.2.0"
sentry-sdk = "^1.0.0"
djangorestframework-simplejwt = "^4.6.0"
munch = "^2.5.0"
Pygments = "^2.8.1"
django-markdownx = "^3.0.1"
model-bakery = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
